+++ sunpy/net/tests/test_vso.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # Author: Florian Mayer <florian.mayer@bitsrc.org>
 
-#pylint: disable=W0613
+# pylint: disable=W0613
 
 from __future__ import absolute_import
 
@@ -93,11 +93,9 @@ def test_complexattr_create(client):
 
 def test_complexattr_and_duplicate():
     attr = va.Time((2011, 1, 1), (2011, 1, 1, 1))
-    pytest.raises(TypeError,
-                  lambda: attr & va.Time((2011, 2, 1), (2011, 2, 1, 1)))
+    pytest.raises(TypeError, lambda: attr & va.Time((2011, 2, 1), (2011, 2, 1, 1)))
     attr |= va.Source('foo')
-    pytest.raises(TypeError,
-                  lambda: attr & va.Time((2011, 2, 1), (2011, 2, 1, 1)))
+    pytest.raises(TypeError, lambda: attr & va.Time((2011, 2, 1), (2011, 2, 1, 1)))
 
 
 def test_complexattr_or_eq():
@@ -110,8 +108,7 @@ def test_complexattr_or_eq():
 def test_attror_and():
     attr = va.Instrument('foo') | va.Instrument('bar')
     one = attr & va.Source('bar')
-    other = ((va.Instrument('foo') & va.Source('bar')) |
-             (va.Instrument('bar') & va.Source('bar')))
+    other = ((va.Instrument('foo') & va.Source('bar')) | (va.Instrument('bar') & va.Source('bar')))
     assert one == other
 
 
@@ -128,14 +125,9 @@ def test_wave_inputQuantity():
 def test_wave_toangstrom():
     # TODO: this test should test that inputs are in any of spectral units
     # more than just converted to Angstroms.
-    frequency = [(1, 1 * u.Hz),
-                 (1e3, 1 * u.kHz),
-                 (1e6, 1 * u.MHz),
-                 (1e9, 1 * u.GHz)]
+    frequency = [(1, 1 * u.Hz), (1e3, 1 * u.kHz), (1e6, 1 * u.MHz), (1e9, 1 * u.GHz)]
 
-    energy = [(1, 1 * u.eV),
-              (1e3, 1 * u.keV),
-              (1e6, 1 * u.MeV)]
+    energy = [(1, 1 * u.eV), (1e3, 1 * u.keV), (1e6, 1 * u.MeV)]
 
     for factor, unit in energy:
         w = va.Wavelength((62 / factor) * unit, (62 / factor) * unit)
@@ -157,32 +149,41 @@ def test_wave_toangstrom():
 
     with pytest.raises(u.UnitsError) as excinfo:
         va.Wavelength(10 * u.g, 23 * u.g)
-    assert 'This unit is not convertable to any of [Unit("Angstrom"), Unit("kHz"), Unit("keV")]' in str(excinfo)
+    assert ('This unit is not convertable to any of '
+            '[Unit("Angstrom"), Unit("kHz"), Unit("keV")]') in str(excinfo)
 
 
 def test_time_xor():
     one = va.Time((2010, 1, 1), (2010, 1, 2))
     a = one ^ va.Time((2010, 1, 1, 1), (2010, 1, 1, 2))
 
-    assert a == attr.AttrOr([va.Time((2010, 1, 1), (2010, 1, 1, 1)), va.Time(
-        (2010, 1, 1, 2), (2010, 1, 2))])
+    assert a == attr.AttrOr(
+        [va.Time((2010, 1, 1), (2010, 1, 1, 1)),
+         va.Time((2010, 1, 1, 2), (2010, 1, 2))])
 
     a ^= va.Time((2010, 1, 1, 4), (2010, 1, 1, 5))
-    assert a == attr.AttrOr([va.Time((2010, 1, 1), (2010, 1, 1, 1)), va.Time(
-        (2010, 1, 1, 2), (2010, 1, 1, 4)), va.Time((2010, 1, 1, 5),
-                                                   (2010, 1, 2))])
+    assert a == attr.AttrOr([
+        va.Time((2010, 1, 1), (2010, 1, 1, 1)),
+        va.Time((2010, 1, 1, 2), (2010, 1, 1, 4)),
+        va.Time((2010, 1, 1, 5), (2010, 1, 2))
+    ])
 
 
 def test_wave_xor():
     one = va.Wavelength(0 * u.AA, 1000 * u.AA)
     a = one ^ va.Wavelength(200 * u.AA, 400 * u.AA)
 
-    assert a == attr.AttrOr([va.Wavelength(0 * u.AA, 200 * u.AA), va.Wavelength(400 * u.AA, 1000 * u.AA)])
+    assert a == attr.AttrOr(
+        [va.Wavelength(0 * u.AA, 200 * u.AA),
+         va.Wavelength(400 * u.AA, 1000 * u.AA)])
 
     a ^= va.Wavelength(600 * u.AA, 800 * u.AA)
 
-    assert a == attr.AttrOr(
-        [va.Wavelength(0 * u.AA, 200 * u.AA), va.Wavelength(400 * u.AA, 600 * u.AA), va.Wavelength(800 * u.AA, 1000 * u.AA)])
+    assert a == attr.AttrOr([
+        va.Wavelength(0 * u.AA, 200 * u.AA),
+        va.Wavelength(400 * u.AA, 600 * u.AA),
+        va.Wavelength(800 * u.AA, 1000 * u.AA)
+    ])
 
 
 def test_err_dummyattr_create():
@@ -202,13 +203,18 @@ def test_wave_repr():
     assert repr(wav) == "<Wavelength(12.0, 16.0, 'Angstrom')>"
     assert repr(moarwav) == "<Wavelength(12.0, 15.0, 'Angstrom')>"
 
+
 def test_str():
     qr = QueryResponse([])
-    assert str(qr) == 'Start Time End Time Source Instrument Type\n---------- -------- ------ ---------- ----'
+    assert str(
+        qr
+    ) == 'Start Time End Time Source Instrument Type\n---------- -------- ------ ---------- ----'
+
 
 def test_repr():
     qr = QueryResponse([])
     assert "Start Time End Time  Source Instrument   Type" in repr(qr)
+
 
 @pytest.mark.online
 def test_path(client):
@@ -216,10 +222,9 @@ def test_path(client):
     Test that '{file}' is automatically appended to the end of a custom path if
     it is not specified.
     """
-    qr = client.query(va.Time('2011-06-07 06:33',
-                              '2011-06-07 06:33:08'),
-                      va.Instrument('aia'),
-                      va.Wavelength(171*u.AA))
+    qr = client.query(
+        va.Time('2011-06-07 06:33', '2011-06-07 06:33:08'),
+        va.Instrument('aia'), va.Wavelength(171 * u.AA))
     tmp_dir = tempfile.mkdtemp()
     files = client.get(qr, path=tmp_dir).wait(progress=False)
 
@@ -229,3 +234,11 @@ def test_path(client):
     # practical way to determine what it should be in this test, so we just
     # put it here.
     assert "aia_lev1_171a_2011_06_07t06_33_02_77z_image_lev1.fits" in files[0]
+
+
+def test_non_str_instrument():
+    # Sanity Check
+    assert isinstance(va.Instrument("lyra"), va.Instrument)
+
+    with pytest.raises(ValueError):
+        va.Instrument(1234)